+++ C# Practice/OOP 1/Student.cs
@@ -1,0 +1,20 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OOP_1
+{
+    internal class Student: Person
+    {
+        public void Study()
+        {
+            Console.WriteLine("I'm studying");
+        }
+        public void ShowAge()
+        {
+            Console.WriteLine("My age is: " + age + " years old");
+        }
+    }
+}
