+++ C# Practice/OOP 1/Teacher.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OOP_1
+{
+    internal class Teacher: Person
+    {
+        public void Explain()
+        {
+            Console.WriteLine("I'm explaining");
+        }
+    }
+}
