+++ C# Practice/OOP 2/BigPhotoBook.cs
@@ -1,0 +1,17 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OOP_2
+{
+    internal class BigPhotoBook: PhotoBook
+    {
+        public BigPhotoBook()
+        {
+            NumPages = 64;
+        }
+    }
+}
+
