+++ C# Practice/OOP 2/PhotoBookTest.cs
@@ -43,36 +43,4 @@ namespace OOP_2
 
         }
     }
-
-    public class PhotoBook
-    {
-        private int numPages;
-        public PhotoBook()
-        {
-            numPages = 16;
-        }
-        public PhotoBook(int numPages)
-        {
-            this.numPages = numPages;
-        }
-
-        public int GetNumberPages() 
-        { 
-            return numPages; 
-        }
-
-        public int NumPages
-        {
-            get { return numPages; }
-            set { numPages = value; }
-        }
-    }
-
-    class BigPhotoBook: PhotoBook
-    {
-        public BigPhotoBook()
-        {
-            NumPages = 64; 
-        }
-    }
 }
