+++ C# Practice/OOP 3/IJob.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OOP_3
+{
+    internal interface IJob
+    {
+        void DisplayName();
+        void DisplayJob();
+    }
+}
